Extract Storage class from Bob.java
